requirements.txt
numpy~=1.21.4
PyYAML~=6.0
pytest~=7.1.2
